fixed orient unit tests for relations, composite keys

+++ src/test/java/org/normandra/orientdb/BaseOrientTest.java
@@ -3,7 +3,6 @@ package org.normandra.orientdb;
 import com.orientechnologies.orient.core.db.document.ODatabaseDocumentTx;
 import org.apache.commons.io.FileUtils;
 import org.apache.log4j.BasicConfigurator;
-import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.normandra.DatabaseConstruction;
 
@@ -17,7 +16,7 @@ import java.io.File;
  */
 public class BaseOrientTest
 {
-    public static final DatabaseConstruction construction = DatabaseConstruction.UPDATE;
+    public static final DatabaseConstruction construction = DatabaseConstruction.RECREATE;
 
     public static final File dir = new File("target/embeddedOrientDB");
 
@@ -29,18 +28,13 @@ public class BaseOrientTest
 
 
     @BeforeClass
-    public static void createLocal() throws Exception
+    public static void setupOrientDB() throws Exception
     {
         BasicConfigurator.configure();
-        FileUtils.deleteDirectory(dir);
-        final ODatabaseDocumentTx db = new ODatabaseDocumentTx(path).create();
-        db.close();
-    }
-
-
-    @AfterClass
-    public static void detroyLocal() throws Exception
-    {
-        FileUtils.deleteDirectory(dir);
+        if(dir.exists())
+        {
+            FileUtils.deleteDirectory(dir);
+        }
+        new ODatabaseDocumentTx(path).create().close();
     }
 }
